Replace existing `MonitoringMessageHandler` with the new DSL
